+++ src/components/Home/home.css
@@ -1,0 +1,13 @@
+.LoginForm {
+    display: flex;
+    justify-content: center;
+}
+h1 {
+    display:flex;
+    justify-content: center;
+    
+}
+#signUp {
+  display:flex;
+    position: right;
+}
