+++ src/components/loginForm/LoginForm.js
@@ -3,6 +3,7 @@ import Spinner from "react-spinkit";
 import { withAsyncAction } from "../../redux/HOCs";
 import "./LoginForm.css";
 
+
 class LoginForm extends React.Component {
   constructor(props) {
     super(props)
