+++ src/components/menu/Menu.css
@@ -1,12 +1,16 @@
 .Menu {
-  background-color: lightskyblue;
+  background-color: #043123;
   display: flex;
   justify-content: space-between;
   align-items: center;
   padding: 0 1em;
   border-radius: 1em;
-}
+  
 
+}
+h1 {
+  color: darkred;
+}
 #menu-links {
   display: flex;
   padding-left: 1em;
@@ -14,7 +18,7 @@
 
 #menu-links a {
   padding-right: 0.5em;
-  color: darkblue;
+  color: darkred;
   text-decoration: underline;
   cursor: pointer;
 }
