+++ src/components/profile/Profile.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Menu from "../components/menu/Menu";
-import { userIsAuthenticated } from "../redux/HOCs";
+import Menu from "../menu/Menu";
+import { userIsAuthenticated } from "../../redux/HOCs";
 
 class Profile extends React.Component {
   render() {
