+++ src/components/registrationForm/RegistrationForm.css
@@ -1,5 +1,5 @@
 #registration-form {
     display: flex;
     flex-direction: column;
-    max-width: 20em;
+    max-width: 120em;
   }
