+++ src/components/registrationForm/RegistrationForm.js
@@ -2,7 +2,7 @@ import React from "react";
 import Spinner from "react-spinkit";
 //import { withAsyncAction } from "../../redux/HOCs";
 import "./RegistrationForm.css";
-import AlmostTwitterService from "../../AlmostTwitterService";
+import AlmostTwitterService from "../../almostTwitterService";
 
 class RegistrationForm extends React.Component {
   constructor(props) {
